refactor: LatestActivity section
